cvic08 uprava čísel na menšie :)

+++ src/com/appslab/Main.java
@@ -1,9 +1,8 @@
 package com.appslab;
-
 import java.util.ArrayList;
-import java.util.Random;
+import java.util.Collection;
 import java.util.Collections;
-
+import java.util.Random;
 public class Main {
     public static void main(String[] args) {
         System.out.println(sum());
@@ -14,8 +13,8 @@ public class Main {
         var arList = new ArrayList<Integer>();
         for (int i = 0; i < 10; i++)
             arList.add(rand.nextInt());
-        System.out.println(Collections.max(arList));
-        System.out.println(Collections.min(arList));
-        return (Collections.max(arList) - Collections.min(arList));
+        System.out.println(Collections.max(arList).byteValue());
+        System.out.println(Collections.min(arList).byteValue());
+        return (Collections.max(arList).byteValue() - Collections.min(arList).byteValue());
     }
 }
